fix(database): use ANY instead of IN for large inputs to avoid limits
PostgreSQL fails when there are > 32767 binded variables.
This commit changes ConditionBuilder "in" method to use "column = ANY(...)" with a single variable instead of "column IN (...)" with 1000s
variables.
It is applied when more than 100 values are passed.
It is required to pass a column type to "in" method, otherwise IN will be always used.