fix(schema-migrations): track index names to avoid collisions
When creating an index for FK (or composed PK on junction table), node-pg-migrate automatically generates index name based on table/columns.
When table/col is renamed, index name is not changed. When a table/col with same name is later created, it name collides with earlier index name.
This issue only affects indexes, constraint names must be unique within a table.

First idea was to track renames and rename the indexes, but it was quite complex. This just tracks taken index names.